Updated documentation to work with gh-pages
- Added index.html to docs folder that redirects to real home page
- Also updated docs
- Added link to docs in README

There was a massive problem with jekyll (a library that github automatically applies to gh-pages sites) where it would not host any folders that began with an underscore.
Initially this was fine as I could change the name of the build folder from '_build' to 'build', which worked fine. However it would not get any css or js files as they were stored in a folder called '_static'.

There was meant to be a quick fix by just adding an empty file called '.nojekyll'. However, this rarely did anything to help, but when it did it would fix it for maybe 5 minutes then it would stop working again. I tried it in literally every location and it would not work. I even tried the sphynx extension (sphinx.ext.githubpages or something) that automatically adds the file to where it needs to be on build, but that didn't work either.

I managed to find a separately built extension that would rename the _static folder but that pointed to the .nojekyll file above and also was written in Python2, so was no use anyway.

After like 2 hours I was fed up and decided to just write a script that would rename the folder to not have an underscore, then just go through every html file created and update any mention of the folder... And this worked. It's not the correct way to do it, but I don't really care at this point because it works and it's reliable.

This is all in one commit, but I should not that this is actually a squash of about 20 commits just testing out different possible solutions to this unnecessary problem.